+++ shape/dot_test.go
@@ -1,0 +1,8 @@
+package shape
+
+import "testing"
+
+func TestDot(t *testing.T) {
+	c := NewDot(24)
+	testShape(t, c)
+}